Converted backend to use doubles over singles (float). it's technically smarter to use floats, because precision isn't really an issue, but the default for floating point numbers is double and as a result it's much easier to work with. may consider do so with bytes and shorts in the future as well.
Auto-Implemented the various update/restore texts for all transformations. In the process of doing so for all remaining tf effects.

Adjusted wearables and armors to handle thing that were previously hard-coded string comparisons in the vanilla. aside from some very unique interactions, it is now possible to use a common variable or interface for these checks, making them (more) future-proof.

Weapons partially implemented; will likely be adjusted to work with new combat system things when implemented. Currently excluded.

+++ CoC.Backend/BodyParts/Ass.cs
@@ -353,7 +353,7 @@ namespace CoC.Backend.BodyParts
 		//Alias these in the creature class, adding the relevant features not in Ass itself (knockup, orgasm)
 		#region Unique Functions
 
-		internal bool PenetrateAsshole(ushort penetratorArea, float knotArea, float cumAmount, bool takeAnalVirginity, bool reachOrgasm, bool sourceIsSelf)
+		internal bool PenetrateAsshole(ushort penetratorArea, double knotArea, double cumAmount, bool takeAnalVirginity, bool reachOrgasm, bool sourceIsSelf)
 		{
 			totalPenetrateCount++;
 			if (sourceIsSelf)
@@ -397,7 +397,7 @@ namespace CoC.Backend.BodyParts
 			totalAnalBirths++;
 		}
 
-		private void HandleStretching(ushort penetratorArea, float knotArea)
+		private void HandleStretching(ushort penetratorArea, double knotArea)
 		{
 			ushort capacity = AnalCapacity();
 			//don't have to worry about overflow, as +1 or +2 will never overflow our artificial max.
